requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
args==0.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
auto-py-to-exe==2.36.0
beautifulsoup4==4.12.2
bidict==0.22.1
bitsandbytes==0.40.2
blinker==1.6.2
boto3==1.28.25
botocore==1.31.25
bottle==0.12.25
bottle-websocket==0.2.9
CacheControl==0.13.1
cachetools==5.3.0
cattrs==1.5.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
clint==0.5.1
cloudpickle==2.2.1
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cryptography==41.0.7
cycler==0.10.0
dataclasses-json==0.5.7
decorator==5.1.1
discord-webhook==1.3.1
discord.py==2.2.3
docopt==0.6.2
EasyProcess==1.1
Eel==0.16.0
einops==0.6.1
encodec==0.1.1
entrypoint2==1.1
fairscale==0.4.13
fake-useragent==1.1.3
filelock==3.11.0
fire==0.5.0
firebase-admin==6.4.0
Flask==2.3.2
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
flask-sock==0.6.0
Flask-SocketIO==5.3.5
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.6.0
funcy==2.0
future==0.18.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gpt-index==0.5.18
gptcache==0.1.16
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.16.4
idna==2.10
imutils==0.5.4
inputimeout==1.0.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
langchain==0.0.142
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mathparse==0.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mlagents==0.28.0
mlagents-envs==0.28.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
noisereduce==2.0.1
numba==0.57.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
patool==1.12
pefile==2023.2.7
Pillow==9.5.0
Pint==0.20.1
pipreqs==0.4.13
playsound==1.2.2
pooch==1.6.0
proto-plus==1.22.3
protobuf==4.22.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.8.0
pymongo==3.13.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==2.4.7
pypiwin32==223
pyqtcolor==2.8
pyspnego==0.9.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.1
python-socketio==5.8.0
pyttsx3==2.90
pytz==2023.3
pyunpack==0.3
pywin32==306
pywin32-ctypes==0.2.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.2
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.3
sahi==0.11.14
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
shapely==2.0.1
simple-websocket==0.10.1
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
suno-bark @ git+https://github.com/suno-ai/bark.git@cb89688307c28cbd2d8bbfc78e534b9812673a26
supervision==0.13.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tiktoken==0.3.3
tldextract==3.4.4
tokenizer==3.4.2
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.32.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.153
uritemplate==4.1.1
urllib3==1.26.15
uTidylib==0.9
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.4
wget==3.2
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.8.2
yolov5==7.0.12
zope.event==4.6
zope.interface==6.0
